🔨 - move state to <Minesweeper /> component
