+++ src/components/minesweeper/index.js
@@ -1,5 +1,19 @@
+import { linkEvent } from 'inferno';
+import Component from 'inferno-component';
 import Board from '../board';
-import { flaggedTiles, lost, won, sweeping } from '../../lib/minesweeper';
+import {
+  board,
+  lost,
+  won,
+  flagTile,
+  gameWon,
+  swept,
+  flagged,
+  iterativeSweep,
+  safe,
+  flaggedTiles,
+  sweeping,
+} from '../../lib/minesweeper';
 import './minesweeper.css';
 
 function emotion(game) {
@@ -15,7 +29,112 @@ function emotion(game) {
   }
 };
 
-export default function Minesweeper({ tiles, threats, cols, onTileClick, game, mines, onReset }) {
+function isRightClick(type) {
+  return type === 2;
+}
+
+function handleTileClick(instance) {
+  return function(pos, event) {
+    const { 
+      tiles,
+      game,
+      mines,
+      threats,
+      cols,
+     } = instance.state;
+    if (game & lost || game & won) return;
+    if (tiles[pos] & flagged) return;
+    if (isRightClick(event.button)) {
+      if (tiles[pos] & swept) return;
+      const newTiles = flagTile(pos, tiles);
+      const wonGame = gameWon(newTiles, mines);
+      instance.setState(function(prevState) {
+        return {
+          tiles: newTiles,
+          game: wonGame ? won : prevState.game,
+        }
+      });
+      return false;
+    }
+    const newTiles = iterativeSweep(pos, tiles, threats, cols);
+    const gameOver = !(safe(newTiles));
+    if (gameOver) {
+      instance.setState({
+        tiles: newTiles,
+        game: lost,
+      });
+      return false;
+    };
+    const wonGame = gameWon(newTiles, mines);
+    if (wonGame) {
+      instance.setState({
+        tiles: newTiles,
+        game: won,
+      });
+      return true;
+    }
+    instance.setState({
+      tiles: newTiles,
+    });
+    return false;   
+  }
+}
+
+function buildBoard(instance) {
+  instance.setState(function(prevState) {
+    return {
+      ...board(prevState.rows, prevState.cols, prevState.mines),
+    }
+  });
+};
+
+export default class Minesweeper extends Component {
+  constructor(props) {
+    super(props);
+    if (props.tiles) {
+      this.state = {
+        ...board(props.tiles),
+      }
+    }
+    if (props.rows && props.cols) {
+      this.state = {
+        ...board(props.row, props.cols, props.mines),
+      }
+    }
+  }
+
+  state = {
+    ...board(9,9,10),
+  };
+
+  componentDidMount() {
+    const game = document.getElementById('minesweeper');
+    game.oncontextmenu = function(event) {
+      event.preventDefault();
+      event.stopPropagation();
+      return false;
+    };
+  }
+
+  render() {
+    const { tiles, threats, cols,  game, mines } = this.state;
+    const { data: onTileClick } = linkEvent(handleTileClick(this));
+    return (
+      <div style={{ width: `${(cols * 16) + 40}px`}} className="game-container" id="minesweeper">
+        <div className="control-panel" style={{ width: `${(cols * 16) + 2}px`}}>
+          <div>💣 {mines - flaggedTiles(tiles)}</div>
+          <div><span onClick={linkEvent(this, buildBoard)}>{ emotion(game) }</span></div>
+          <div>⏱ {'000'}</div>
+        </div>
+        <div style={{ width: `${(cols * 16) + 2}px`}} className='board' noNormalize hasNonKeyedChildren>
+          {Board({ tiles, threats, cols, onTileClick, game })}
+        </div>
+      </div>
+    );
+  }
+}
+
+/*export default function Minesweeper({ tiles, threats, cols, onTileClick, game, mines, onReset }) {
   return (
     <div style={{ width: `${(cols * 16) + 40}px`}} className="game-container" id="minesweeper">
       <div className="control-panel" style={{ width: `${(cols * 16) + 2}px`}}>
@@ -28,4 +147,4 @@ export default function Minesweeper({ tiles, threats, cols, onTileClick, game, m
       </div>
     </div>
   );
-}
+}*/
